fix(db): make the seeder and generated schema work on SQLite
db seed failed immediately on SQLite because the seed bookkeeping DDL was
hardcoded PostgreSQL. Fixing that surfaced four further blockers, found by
running the whole seed registry against a migrated SQLite database.

Seed bookkeeping DDL now lives in per-dialect .sql files under
seeder/schema/, embedded with go:embed and split on --bun:split, so the SQL
stays reviewable instead of sitting in Go string literals. Statements execute
one at a time, which SQLite drivers require and which names the failing
statement when something breaks.

Model defaults: the jsonb casts in bun default tags ('{}'::jsonb) emitted a
'::' token SQLite cannot parse. The casts are redundant on PostgreSQL too,
since a string literal assignment-casts to jsonb, so they are simply dropped.

Converter fixes:
- Array columns map to TEXT holding JSON, so the PostgreSQL empty-array literal
  '{}' has to become '[]'. As an empty JSON object it failed to unmarshal into
  a Go slice.
- numeric/decimal map to REAL rather than NUMERIC. NUMERIC affinity demotes
  integral values to INTEGER, which then refuses to scan into a Go float64.
  Precision is unchanged; SQLite has no exact decimal either way.
- DROP COLUMN is emitted again instead of being skipped wholesale. Skipping it
  left columns PostgreSQL had dropped behind as NOT NULL, which broke inserts.
  SQLite only refuses the drop when an index or constraint still references the
  column, so the converter now tracks emitted indexes and constraint columns,
  drops blocking indexes first, and skips only genuinely blocked drops.

All 1518 translated statements apply cleanly, up from 1471, and the seed run
reaches 21 of 25 seeds. The remaining blocker is an upstream bun bug, reported
separately.

+++ services/tms/internal/infrastructure/sqlite/migrations/20250608193352_dedicated_lane_config.tx.up.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS "pattern_configs"(
     "enabled" INTEGER NOT NULL DEFAULT 1,
     "min_frequency" INTEGER NOT NULL DEFAULT 3,
     "analysis_window_days" INTEGER NOT NULL DEFAULT 90,
-    "min_confidence_score" NUMERIC NOT NULL DEFAULT 0.7,
+    "min_confidence_score" REAL NOT NULL DEFAULT 0.7,
     "suggestion_ttl_days" INTEGER NOT NULL DEFAULT 30,
     "require_exact_match" INTEGER NOT NULL DEFAULT 0,
     "weight_recent_shipments" INTEGER NOT NULL DEFAULT 1,
