+++ services/tms/internal/infrastructure/sqlite/migrations/20260720000000_order_charges.tx.up.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS "order_charges"(
     "organization_id" TEXT NOT NULL,
     "order_id" TEXT NOT NULL,
     "description" TEXT NOT NULL,
-    "amount" NUMERIC NOT NULL DEFAULT 0,
+    "amount" REAL NOT NULL DEFAULT 0,
     "version" INTEGER NOT NULL DEFAULT 0,
     "created_at" INTEGER NOT NULL DEFAULT (unixepoch()),
     "updated_at" INTEGER NOT NULL DEFAULT (unixepoch()),
